+++ src/main/java/com/example/CargoTracking/model/InternationalShipment.java
@@ -75,7 +75,7 @@ public class InternationalShipment {
     private boolean escalationFlagThree;
     private boolean activeStatus;
     private int trip;
-    private Double damage;
+    private Integer damage;
     private String damageAwbs;
 
     @ManyToOne
